Separate SeqMapperNotifier
