+++ src/projects/hpcspades/common/alignment/CMakeLists.txt
@@ -1,12 +1,12 @@
 ############################################################################
 # Copyright (c) 2021 Saint Petersburg State University
-# Copyright (c) 2011-2014 Saint Petersburg Academic University
 # All Rights Reserved
 # See file LICENSE for details.
 ############################################################################
 
-project(common_modules_mpi CXX)
+project(alignment_mpi CXX)
 
-add_subdirectory(paired_info)
-add_subdirectory(alignment)
-add_subdirectory(stages)
+add_library(alignment_mpi STATIC
+            sequence_mapper_notifier_mpi.cpp)
+
+target_link_libraries(alignment_mpi modules)
