+++ src/projects/hpcspades/distance_estimation_mpi.hpp
@@ -9,6 +9,8 @@
 
 #include "common/pipeline/mpi_stage.hpp"
 #include "projects/spades/distance_estimation.hpp"
+#include "common/alignment/sequence_mapper_notifier_mpi.hpp"
+
 
 namespace debruijn_graph {
     class DistanceEstimationMPI : public DistanceEstimationBase, public spades::MPIAssemblyStage {
