+++ src/projects/hpcspades/gap_closer_mpi.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "projects/spades/gap_closer.hpp"
-#include "alignment/sequence_mapper_notifier.hpp"
+#include "common/alignment/sequence_mapper_notifier_mpi.hpp"
 #include "pipeline/mpi_stage.hpp"
 #include "io/reads/io_helper.hpp"
