+++ src/projects/hpcspades/mismatch_correction_mpi.hpp
@@ -7,6 +7,7 @@
 #pragma once
 
 #include "projects/spades/mismatch_correction.hpp"
+#include "common/alignment/sequence_mapper_notifier_mpi.hpp"
 #include "pipeline/mpi_stage.hpp"
 #include "pipeline/graph_pack_helpers.h"
